requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.4
django-ninja==1.1.0
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.7.1
pydantic_core==2.18.2
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0
